refactor: continue simplifying TODO use case usage
