+++ boilerplates/drizzle/package.json
@@ -42,9 +42,6 @@
     },
     "./database/seed": {
       "types": "./dist/types/database/seed.d.ts"
-    },
-    "./server/create-todo-handler": {
-      "types": "./dist/types/server/create-todo-handler.d.ts"
     }
   },
   "typesVersions": {
@@ -60,9 +57,6 @@
       ],
       "database/seed": [
         "./dist/types/database/seed.d.ts"
-      ],
-      "server/create-todo-handler": [
-        "./dist/types/server/create-todo-handler.d.ts"
       ]
     }
   }
