+++ boilerplates/express/files/express-entry.ts
@@ -3,13 +3,13 @@ import "dotenv/config";
 import { dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 import { authjsHandler, authjsSessionMiddleware } from "@batijs/authjs/server/authjs-handler";
-import { createTodoHandler } from "@batijs/drizzle/server/create-todo-handler";
 import {
   firebaseAuthLoginHandler,
   firebaseAuthLogoutHandler,
   firebaseAuthMiddleware,
 } from "@batijs/firebase-auth/server/firebase-auth-middleware";
 import { vikeHandler } from "@batijs/shared-server/server/vike-handler";
+import { createTodoHandler } from "@batijs/shared-todo/server/create-todo-handler";
 import { telefuncHandler } from "@batijs/telefunc/server/telefunc-handler";
 import { appRouter } from "@batijs/trpc/trpc/server";
 import * as trpcExpress from "@trpc/server/adapters/express";
@@ -117,7 +117,7 @@ async function startServer() {
     app.post("/_telefunc", handlerAdapter(telefuncHandler));
   }
 
-  if (BATI.has("drizzle") && !(BATI.has("telefunc") || BATI.has("trpc"))) {
+  if (!BATI.has("telefunc") && !BATI.has("trpc")) {
     app.post("/api/todo/create", handlerAdapter(createTodoHandler));
   }
