+++ boilerplates/h3/files/h3-entry.ts
@@ -4,13 +4,13 @@ import { createServer, type IncomingMessage, type ServerResponse } from "node:ht
 import { dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 import { authjsHandler, authjsSessionMiddleware } from "@batijs/authjs/server/authjs-handler";
-import { createTodoHandler } from "@batijs/drizzle/server/create-todo-handler";
 import {
   firebaseAuthLoginHandler,
   firebaseAuthLogoutHandler,
   firebaseAuthMiddleware,
 } from "@batijs/firebase-auth/server/firebase-auth-middleware";
 import { vikeHandler } from "@batijs/shared-server/server/vike-handler";
+import { createTodoHandler } from "@batijs/shared-todo/server/create-todo-handler";
 import { telefuncHandler } from "@batijs/telefunc/server/telefunc-handler";
 import { appRouter } from "@batijs/trpc/trpc/server";
 import installCrypto from "@hattip/polyfills/crypto";
@@ -126,7 +126,7 @@ async function startServer() {
     router.post("/_telefunc", fromWebHandler(telefuncHandler));
   }
 
-  if (BATI.has("drizzle") && !(BATI.has("telefunc") || BATI.has("trpc"))) {
+  if (!BATI.has("telefunc") && !BATI.has("trpc")) {
     router.post("/api/todo/create", fromWebHandler(createTodoHandler));
   }
