+++ boilerplates/h3/package.json
@@ -18,6 +18,7 @@
     "@batijs/drizzle": "workspace:*",
     "@batijs/firebase-auth": "workspace:*",
     "@batijs/shared-server": "workspace:^",
+    "@batijs/shared-todo": "workspace:^",
     "@batijs/telefunc": "workspace:^",
     "@batijs/trpc": "workspace:^",
     "@hattip/polyfills": "^0.0.45",
