+++ boilerplates/hattip/files/hattip-entry.ts
@@ -1,13 +1,13 @@
 // BATI.has("auth0")
 import "dotenv/config";
 import { authjsHandler, authjsSessionMiddleware } from "@batijs/authjs/server/authjs-handler";
-import { createTodoHandler } from "@batijs/drizzle/server/create-todo-handler";
 import {
   firebaseAuthLoginHandler,
   firebaseAuthLogoutHandler,
   firebaseAuthMiddleware,
 } from "@batijs/firebase-auth/server/firebase-auth-middleware";
 import { vikeHandler } from "@batijs/shared-server/server/vike-handler";
+import { createTodoHandler } from "@batijs/shared-todo/server/create-todo-handler";
 import { telefuncHandler } from "@batijs/telefunc/server/telefunc-handler";
 import { appRouter } from "@batijs/trpc/trpc/server";
 import type { HattipHandler } from "@hattip/core";
@@ -76,7 +76,7 @@ if (BATI.has("firebase-auth")) {
   router.post("/api/sessionLogout", handlerAdapter(firebaseAuthLogoutHandler));
 }
 
-if (BATI.has("drizzle") && !(BATI.has("telefunc") || BATI.has("trpc"))) {
+if (!BATI.has("telefunc") && !BATI.has("trpc")) {
   router.post("/api/todo/create", handlerAdapter(createTodoHandler));
 }
