+++ boilerplates/hono/files/hono-entry.ts
@@ -1,13 +1,13 @@
 // BATI.has("auth0")
 import "dotenv/config";
 import { authjsHandler, authjsSessionMiddleware } from "@batijs/authjs/server/authjs-handler";
-import { createTodoHandler } from "@batijs/drizzle/server/create-todo-handler";
 import {
   firebaseAuthLoginHandler,
   firebaseAuthLogoutHandler,
   firebaseAuthMiddleware,
 } from "@batijs/firebase-auth/server/firebase-auth-middleware";
 import { vikeHandler } from "@batijs/shared-server/server/vike-handler";
+import { createTodoHandler } from "@batijs/shared-todo/server/create-todo-handler";
 import { telefuncHandler } from "@batijs/telefunc/server/telefunc-handler";
 import { appRouter } from "@batijs/trpc/trpc/server";
 import { serve } from "@hono/node-server";
@@ -104,7 +104,7 @@ if (BATI.has("telefunc")) {
   app.post("/_telefunc", handlerAdapter(telefuncHandler));
 }
 
-if (BATI.has("drizzle") && !(BATI.has("telefunc") || BATI.has("trpc"))) {
+if (!BATI.has("telefunc") && !BATI.has("trpc")) {
   app.post("/api/todo/create", handlerAdapter(createTodoHandler));
 }
