+++ boilerplates/hono/package.json
@@ -18,6 +18,7 @@
     "@batijs/drizzle": "workspace:*",
     "@batijs/firebase-auth": "workspace:*",
     "@batijs/shared-server": "workspace:^",
+    "@batijs/shared-todo": "workspace:^",
     "@batijs/telefunc": "workspace:^",
     "@batijs/trpc": "workspace:^",
     "@hono/node-server": "^1.11.1",
