+++ boilerplates/react/package.json
@@ -14,7 +14,6 @@
   "devDependencies": {
     "@batijs/compile": "workspace:^",
     "@batijs/drizzle": "workspace:^",
-    "@batijs/shared-db": "workspace:^",
     "@batijs/shared-todo": "workspace:^",
     "@batijs/telefunc": "workspace:^",
     "@batijs/trpc": "workspace:^",
