--- boilerplates/shared-db/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@batijs/shared-db",
-  "private": true,
-  "version": "0.0.1",
-  "description": "",
-  "type": "module",
-  "scripts": {
-    "check-types": "tsc --noEmit",
-    "build": "bati-compile-boilerplate"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "@batijs/compile": "workspace:*",
-    "@types/node": "^18.19.14"
-  },
-  "dependencies": {
-    "@batijs/core": "workspace:*"
-  },
-  "files": [
-    "dist/"
-  ],
-  "bati": {
-    "if": {
-      "flag": {
-        "$in": [
-          "trpc",
-          "telefunc"
-        ]
-      }
-    }
-  },
-  "exports": {
-    "./database/todoItems": {
-      "types": "./dist/types/database/todoItems.d.ts"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "database/todoItems": [
-        "./dist/types/database/todoItems.d.ts"
-      ]
-    }
-  }
-}
