--- boilerplates/shared-db/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": [
-    "../tsconfig.base.json"
-  ]
-}
