+++ boilerplates/shared-todo/files/pages/todo/+data.ts
@@ -1,7 +1,7 @@
 // https://vike.dev/data
 import { db } from "@batijs/drizzle/database/db";
 import { todoTable } from "@batijs/drizzle/database/schema";
-import { todoItems } from "@batijs/shared-db/database/todoItems";
+import { todoItems } from "@batijs/shared-todo/database/todoItems";
 
 export type Data = { text: string }[];
