+++ boilerplates/telefunc/package.json
@@ -14,7 +14,7 @@
   "devDependencies": {
     "@batijs/compile": "workspace:*",
     "@batijs/drizzle": "workspace:*",
-    "@batijs/shared-db": "workspace:*",
+    "@batijs/shared-todo": "workspace:*",
     "@types/node": "^18.19.14",
     "telefunc": "^0.1.72",
     "vike": "^0.4.172",
