+++ boilerplates/trpc/package.json
@@ -14,12 +14,10 @@
   "devDependencies": {
     "@batijs/compile": "workspace:*",
     "@batijs/drizzle": "workspace:*",
-    "@batijs/shared-db": "workspace:*",
-    "@types/better-sqlite3": "^7.6.10",
+    "@batijs/shared-todo": "workspace:*",
     "@trpc/client": "^10.45.2",
     "@trpc/server": "^10.45.2",
     "@types/node": "^18.19.14",
-    "better-sqlite3": "^10.0.0",
     "vite": "^5.2.11"
   },
   "dependencies": {
