+++ packages/tests/tests/FRAMEWORK+SERVER+DATA.spec.ts
@@ -4,15 +4,17 @@ export const matrix = [
   ["solid", "react", "vue"],
   ["express", "h3", "hono", "fastify", "hattip"],
   ["trpc", "telefunc", undefined],
-  "drizzle",
+  ["drizzle", undefined],
   "eslint",
-];
+] as const;
 
 await describeBati(({ test, expect, fetch, testMatch, context, beforeAll }) => {
   beforeAll(async () => {
-    await exec(npmCli, ["run", "drizzle:generate"]);
-    await exec(npmCli, ["run", "drizzle:migrate"]);
-    await exec(npmCli, ["run", "drizzle:seed"]);
+    if (context.flags.includes("drizzle")) {
+      await exec(npmCli, ["run", "drizzle:generate"]);
+      await exec(npmCli, ["run", "drizzle:migrate"]);
+      await exec(npmCli, ["run", "drizzle:seed"]);
+    }
   }, 70000);
 
   test("home", async () => {
@@ -27,7 +29,7 @@ await describeBati(({ test, expect, fetch, testMatch, context, beforeAll }) => {
     expect(await res.text()).not.toContain('{"is404":true}');
   });
 
-  if (!context.flags.includes("telefunc") && !context.flags.includes("trpc"))
+  if (!context.flags.includes("telefunc") && !context.flags.includes("trpc")) {
     test("/api/todo/create", async () => {
       const res = await fetch("/api/todo/create", {
         method: "POST",
@@ -36,25 +38,21 @@ await describeBati(({ test, expect, fetch, testMatch, context, beforeAll }) => {
           "content-type": "application/json",
         },
       });
-      expect(res.status).toBe(201);
-      expect(await res.json()).toEqual({
-        message: "New Todo Created",
-        result: { changes: 1, lastInsertRowid: 3 },
-      });
+      expect(res.status).toBe(200);
     });
+  }
 
   testMatch<typeof matrix>("onNewTodo", {
     telefunc: async () => {
       const res = await fetch("/_telefunc", {
         method: "POST",
         body: JSON.stringify({
-          file: "/components/TodoList.telefunc.ts",
+          file: "/pages/todo/TodoList.telefunc.ts",
           name: "onNewTodo",
           args: [{ text: "test" }],
         }),
       });
       expect(res.status).toBe(200);
-      expect(await res.json()).toEqual({ ret: { changes: 1, lastInsertRowid: 3 } });
     },
     trpc: async () => {
       const res = await fetch("/api/trpc/onNewTodo", {
@@ -65,14 +63,6 @@ await describeBati(({ test, expect, fetch, testMatch, context, beforeAll }) => {
         },
       });
       expect(res.status).toBe(200);
-      expect(await res.json()).toEqual({
-        result: {
-          data: {
-            changes: 1,
-            lastInsertRowid: 3,
-          },
-        },
-      });
     },
   });
 });
