--- packages/tests/tests/FRAMEWORK+SERVER+telefunc.spec.ts
@@ -1,30 +1,0 @@
-import { describeBati } from "@batijs/tests-utils";
-
-export const matrix = [["solid", "react", "vue"], ["express", "h3", "hono", "fastify", "hattip"], "telefunc", "eslint"];
-
-await describeBati(({ test, expect, fetch }) => {
-  test("home", async () => {
-    const res = await fetch("/");
-    expect(res.status).toBe(200);
-    expect(await res.text()).not.toContain('{"is404":true}');
-  });
-
-  test("todo", async () => {
-    const res = await fetch("/todo");
-    expect(res.status).toBe(200);
-    expect(await res.text()).not.toContain('{"is404":true}');
-  });
-
-  test("telefunc", async () => {
-    const res = await fetch("/_telefunc", {
-      method: "post",
-    });
-    expect(res.status).toBe(400);
-    expect(await res.text()).not.toContain('{"is404":true}');
-  });
-
-  test("auth/signin", async () => {
-    const res = await fetch("/api/auth/signin");
-    expect(await res.text()).toContain('{"is404":true}');
-  });
-});
